insert runtime checks for object type prior to calling the EventListener
this allows child classes to *not* miss the interception

one of the unobvious difficulties:

Suppose you register A.foo() and B.foo() for interception. Then before each call
to c.foo() we would insert a bytecode equivalent to this sequence:
if (c instanceof A) {
    A a = (A) c;
    EventListener.listenAfoo(a, pc);
}
if (c instanceof B) {
    B b = (B) c;
    EventListenet.listenBfoo(b, pc);
}

listenAfoo must be declared as:
public void listenAfoo(A a, long pc);

so, you see, when methods called 'lock' are intercepted, there is pretty much
code bloating going on.